refactor(error-page): /simplify pass — hoist DISTRICT_SECTIONS + URL builder to config, drop dead guard (#1012)
